[libs] E: Move Binding for renameat()
This commits moves the binding for the renameat() system call to the
syscall crate.
