+++ src/libs/syscall/src/fcntl/bindings/mod.rs
@@ -9,3 +9,4 @@ pub mod fcntl;
 pub mod open;
 pub mod posix_fadvise;
 pub mod posix_fallocate;
+pub mod renameat;
